refactor(GameLogic): Paso todo a double.
Paso los tiempos, distancias, daño, vida, posiciones, etc a double

+++ Server/include/GameLogic/Weapons/p90weapon.h
@@ -11,18 +11,18 @@ class P90Weapon : public Weapon {
 private:
     uint8_t ammo;
     uint8_t actual_ammo;
-    uint8_t damage;
-    uint8_t scope;
-    float damage_reduction_coef;
+    double damage;
+    double scope;
+    double damage_reduction_coef;
 
 public:
-    explicit P90Weapon(uint8_t ammo, uint8_t damage, uint8_t scope, float reduction);
+    explicit P90Weapon(uint8_t ammo, double damage, double scope, double reduction);
 
     bool shoot(
     Position& from,
     int8_t dir,
-    int32_t dim_x,
-    uint16_t time,
+    double dim_x,
+    double time,
     std::map<uint32_t, std::shared_ptr<Soldier>>& soldiers,
     std::map<uint32_t, std::shared_ptr<Zombie>>& zombies) override;
